refactor(shared): mask_dsn 을 utils 로 분리 (lifespan 책임 아님)
DSN password 마스킹은 로깅 안전성 유틸 — lifespan / checkpointer 의 책임이
아니다. shared/utils.py 신설하고 옮김.

- shared/utils.py: mask_dsn (지금 1 함수. 다른 utility 늘면 utils/ 서브패키지로 승격)
- lifespan/checkpointer.py: from dev_team_shared.utils import mask_dsn 으로 전환
- lifespan/__init__.py: mask_dsn re-export 제거
- agents/primary, agents/librarian 의 lifespan_helpers: 같은 re-export 제거
  (실제 server.py 에서 호출하는 곳 없어 영향 없음)
